Set up pre-test deletion and env-specific post test deletion for ProductAttribute mutations

+++ src/ecomm-api-svc/cypress/integration/mutations/productAttribute/updateProductAttribute.spec.ts
@@ -9,6 +9,8 @@ describe('Mutation: updateProductAttribute', () => {
     var values = '';
     var extraIds = [] as SupplementalItemRecord[];
     const mutationName = 'updateProductAttribute';
+    const createName = 'createProductAttribute';
+    const deleteMutName = "deleteProductAttribute";
     const queryName = "productAttributes";
     const itemPath = 'productAttribute';
     const additionalFields = `values {
@@ -25,9 +27,12 @@ describe('Mutation: updateProductAttribute', () => {
             ${additionalFields}
         }
     `;
-    const createName = 'createProductAttribute';
 
+	var deleteItemsAfter = undefined as boolean | undefined;
     before(() => {
+		deleteItemsAfter = Cypress.env("deleteItemsAfter");
+		cy.deleteCypressItems(queryName, deleteMutName);
+        // TODO: Move this to a beforeEach so we're updating a new item each time
         const name = `Cypress ${mutationName} Test`;
         const input = `{name: "${name}", values: [{name: "Cypress PA update test"}]}`;
         cy.createAndGetId(createName, itemPath, input, additionalFields).then((createdItem) => {
@@ -39,11 +44,14 @@ describe('Mutation: updateProductAttribute', () => {
     });
 
     after(() => {
+		if (!deleteItemsAfter) {
+			return;
+		}
         if (id !== "") {
             // Delete any supplemental items we created
             cy.deleteSupplementalItems(extraIds);
             // Delete the item we've been updating
-            cy.deleteItem("deleteProductAttribute", id);
+            cy.deleteItem(deleteMutName, id);
         }
     });
 
@@ -210,7 +218,7 @@ describe('Mutation: updateProductAttribute', () => {
             cy.createAndGetId(createName, itemPath, input, extraInput).then((createdItem) => {
                 assert.exists(createdItem.id);
                 assert.exists(createdItem.customData);
-                extraIds.push({itemId: createdItem.id, deleteName: "deleteProductAttribute", itemName: name, queryName: queryName});
+                extraIds.push({itemId: createdItem.id, deleteName: deleteMutName, itemName: name, queryName: queryName});
                 const newName = `Cypress ${mutationName} CD extra updated`;
                 const newValues = createdItem.values;
                 const newCustomData = {data: `${itemPath} customData`, newDataField: { canDelete: true }};
